Collapse "tool-search-enterprise".
